+++ doc/conf.py
@@ -3,8 +3,8 @@
 This configuration only affects single-package Sphinx documentation builds.
 """
 
-from documenteer.conf.pipelinespkg import *  # noqa
 import lsst.ts.fiberspectrograph  # noqa
+from documenteer.conf.pipelinespkg import *  # noqa
 
 project = "ts_fiberspectrograph"
 html_theme_options["logotext"] = project  # type: ignore # noqa