+++ python/lsst/ts/fiberspectrograph/csc.py
@@ -26,16 +26,13 @@ import io
 import pathlib
 
 import astropy.units as u
-
+from lsst.ts import salobj, utils
 from lsst.ts.idl.enums.FiberSpectrograph import ExposureState
-from lsst.ts import salobj
-from lsst.ts import utils
-from . import constants
-from . import __version__
-from .avsSimulator import AvsSimulator
+
+from . import __version__, constants, dataManager
 from .avsFiberSpectrograph import AvsFiberSpectrograph
+from .avsSimulator import AvsSimulator
 from .config_schema import CONFIG_SCHEMA
-from . import dataManager
 
 
 class FiberSpectrographCsc(salobj.ConfigurableCsc):
@@ -128,6 +125,9 @@ class FiberSpectrographCsc(salobj.ConfigurableCsc):
             s3instance=config.s3instance,
         )
         self.config = config
+        self.image_service_client = utils.ImageNameServiceClient(
+            config.image_service_url, self.salinfo.index, "FiberSpectrograph"
+        )
 
     async def handle_summary_state(self):
         # disabled: connect and send telemetry, but no commands allowed.
@@ -257,6 +257,10 @@ class FiberSpectrographCsc(salobj.ConfigurableCsc):
         upload succeeds.
         """
         hdulist = self.data_manager.make_hdulist(spec_data)
+        image_sequence_array, data = await self.image_service_client.get_next_obs_id(
+            num_images=1
+        )
+        hdulist[0].header["OBSID"] = image_sequence_array[0]
         fileobj = io.BytesIO()
         hdulist.writeto(fileobj)
         fileobj.seek(0)
@@ -270,7 +274,7 @@ class FiberSpectrographCsc(salobj.ConfigurableCsc):
         )
         try:
             await self.s3bucket.upload(fileobj=fileobj, key=key)
-            url = f"s3://{self.s3bucket.name}/{key}"
+            url = f"{self.s3bucket.service_resource.meta.client.meta.endpoint_url}/{self.s3bucket.name}/{key}"
             await self.evt_largeFileObjectAvailable.set_write(
                 url=url, generator=self.generator_name
             )