+++ python/lsst/ts/fiberspectrograph/dataManager.py
@@ -24,11 +24,10 @@ __all__ = ["SpectrographData", "DataManager"]
 import dataclasses
 
 import astropy.io.fits
-import astropy.time
 import astropy.table
+import astropy.time
 import astropy.units as u
 import numpy as np
-
 
 # The version of the FITS file format produced by this class.
 FORMAT_VERSION = 1
@@ -106,8 +105,10 @@ class DataManager:
         # TODO: it would be good to include the dataclass docstrings
         # as comments on each of these, but pydoc can't see them.
         hdr["FORMAT_V"] = FORMAT_VERSION
+        hdr["OBSERVAT"] = "Vera C. Rubin Observatory"
         hdr["INSTRUME"] = self.instrument
         hdr["ORIGIN"] = self.origin
+        hdr["LOCATN"] = (None, "Location of Instrument")
         hdr["DETSIZE"] = data.n_pixels
         hdr["DATE-BEG"] = astropy.time.Time(data.date_begin).tai.fits
         hdr["DATE-END"] = astropy.time.Time(data.date_end).tai.fits